+++ src/util.mli
@@ -1,9 +1,13 @@
-(** Utility functions *)
+module type RANDOM = S.RANDOM
 
-(** Percent encode a string according to RFC 3986, Section 2.1
+module type S = sig
+  
+  (** Percent encode a string according to RFC 3986, Section 2.1
   http://tools.ietf.org/html/rfc3986#section-2.1 *)
-val pct_encode : string -> string
+  val pct_encode : string -> string
 
-type rng = unit -> char
+  val generate_nonce : int -> string
+  
+end
 
-val generate_nonce : rng -> int -> string 
+module Make (Random : RANDOM) : S