+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 0.1.0 / 2014-05-21
 ==================
 
+  * V1.0a api tidied up.
+  * Dependency on Cryptokit removed, default MAC implementation provided
+    by Cryptokit in the oauth-client.unix package.
+
+0.1.0 / 2014-05-21
+==================
+
   * Initial commit, v1.0 of the oauth protocol is supported.